feat(FN-3699): modify text and hide text behind feature flag
