+++ trade-finance-manager-ui/templates/utilisation-reports/partials/premium-payments-html.njk
@@ -39,16 +39,19 @@
         <p class="govuk-body" data-cy="how-to-generate-keying-data-text">
           When payments show as matched, the adjustment data for keying into ACBS will be automatically generated when the 'Generate keying data' button is selected.
         </p>
-
-         <p class="govuk-body" data-cy="how-to-create-record-correction-request-text">
-          If there is an error with the fee record, select the record to create a record correction request to send a query to the bank.
-        </p>
+         {% if isFeeRecordCorrectionFeatureFlagEnabled %}
+          <p class="govuk-body" data-cy="how-to-create-record-correction-request-text">
+            If there is an error with the fee record, select the record to create a record correction request to send a query to the bank.
+          </p>
+        {% endif %}
       {% else %}
         <p class="govuk-body" data-cy="received-payments-text">Received payments are entered against reported fees. When payments show as matched, the adjustment data for keying into ACBS will be automatically generated.</p>
-
-        <p class="govuk-body" data-cy="how-to-create-record-correction-request-text">
-          Errors with a fee record can be addressed and queried with the bank when the 'Create a record correction request' button is selected.
-        </p>
+        
+        {% if isFeeRecordCorrectionFeatureFlagEnabled %}
+          <p class="govuk-body" data-cy="how-to-create-record-correction-request-text">
+            Errors with a fee record can be addressed and queried with the bank through a record correction request.
+          </p>
+        {% endif %}
       {% endif %}
     </div>
